+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Latest changes
 
+* VFS-13176 Added support for the replaced file name in the upload
 * VFS-13143 Added support for changing share's tab in the URL from the embedded Oneprovider
 * VFS-13044 Added search input for harvester spaces
 * VFS-13155 Fixed display of longer membership paths when the display limit is reached